Move auth interfaces to deticated files.
