+++ src/utils/interfaces/auth/IAuthErrorResponse.ts
@@ -1,0 +1,3 @@
+export interface IAuthErrorResponse {
+    detail: string;
+}
