+++ src/utils/interfaces/auth/ILoginCredentials.ts
@@ -1,0 +1,4 @@
+export interface ILoginCredentials {
+    email: string;
+    password: string;
+}
