+++ src/utils/interfaces/auth/IRegisterFormData.ts
@@ -1,0 +1,6 @@
+export interface IRegisterFormData {
+    email: string;
+    password: string;
+    confirmPassword: string;
+    username: string;
+}
