+++ src/utils/interfaces/auth/IRoleBasedRedirect.ts
@@ -1,0 +1,6 @@
+import { Role } from "../../enums";
+
+export interface IRoleBasedRedirect {
+    allowedRoles: Role[];
+    fallbackRoute: string;
+}
