+++ src/utils/interfaces/auth/ITokens.ts
@@ -1,0 +1,4 @@
+export interface ITokens {
+    accessToken: string;
+    refreshToken: string;
+}
