+++ src/utils/interfaces/auth/IUserDetails.ts
@@ -1,0 +1,7 @@
+import { Role } from "../../enums";
+
+export interface IUserDetails {
+    username: string;
+    userId: string;
+    roles: Role[];
+}
