+++ src/utils/interfaces/auth/LoginErrorMessageType.ts
@@ -1,0 +1,4 @@
+export type LoginErrorMessage =
+    | "wrongCredentials"
+    | "serverProblem"
+    | "mustEndWithLtuErrMsg";
