+++ src/utils/interfaces/auth/RegisterErrorMessageType.ts
@@ -1,0 +1,6 @@
+export type RegisterErrorMessage =
+    | "usernameTaken"
+    | "serverProblem"
+    | "passwordsNoMatch"
+    | "mustEndWithLtuErrMsg"
+    | "emailTaken";
